Synced Movement (beginnings)
Made a Sync Music script, with a static variable deltaSample, which is the change in samples of the music per frame. I use this variable now for all of the speeds. Seems to work well, although, movement is a bit shaky. I believe this is because sometimes, deltaSample is =0, and other times, deltaSample = a constant. I will average this delta sample over a couple of them maybe, in order to get smoother movement??